Split out embedder_unittests test cases into GL and non-GL tests (#21386)
